requirements.txt
Django==4.2.5
djangorestframework==3.14
sorl-thumbnail==12.9
Pillow==10.0
pytest==7.4
pytest-django==4.5